Restructure and rename packages to better represent the change
